CR: Move remote introspection to separate module.
